Have main be a test to parse a log file
This will break when not on `tony`! And it should probably be in the
tests, not here.

But it *has* proved that the garbage collecter isn't working quite
right. This won't matter for now, because (a) the log files aren't that
large -- nowhere near the size of this test file -- and (b)
it's *premature optimisation*! Make a ticket, and move on.